requirements.txt
transformers==4.36.2
numpy==1.24.3
torch>=2.2.0
fasttext-wheel==0.9.2
python-dotenv==1.0.1
sentencepiece
accelerate
fastapi==0.109.2
uvicorn==0.27.1
pydantic>=2.0.0
requests>=2.31.0
tqdm>=4.66.1